Filter Panel works on responsive too
- Created a couple new components:

--article-filter.popover: Returns the Popover component with the articles filter panel as content and the filter icon as a trigger.

--compact-search-input: Extracted the "responsive" version of the search input to a new component. It receives the same props as the search-input component. They are parameterized so that they can be used in both Headers.

Also updated some of the structure of how the filters will be passed to the header children.
